remove multer package and implement multer inside platform-fastify
