--- packages/multer/constants.ts
@@ -1,12 +1,0 @@
-export const MULTER_MODULE_OPTIONS = 'MULTER_MODULE_OPTIONS';
-export const MULTER_MODULE_ADAPTER = 'MULTER_MODULE_ADAPTER';
-
-export const multerExceptions = {
-  LIMIT_PART_COUNT: 'Too many parts',
-  LIMIT_FILE_SIZE: 'File too large',
-  LIMIT_FILE_COUNT: 'Too many files',
-  LIMIT_FIELD_KEY: 'Field name too long',
-  LIMIT_FIELD_VALUE: 'Field value too long',
-  LIMIT_FIELD_COUNT: 'Too many fields',
-  LIMIT_UNEXPECTED_FILE: 'Unexpected field',
-};
