--- packages/multer/index.ts
@@ -1,5 +1,0 @@
-export * from './constants';
-export * from './interfaces';
-export * from './multer.module';
-export * from './utils';
-export * from './interceptors';
