--- packages/multer/interceptors/files.interceptor.ts
@@ -1,62 +1,0 @@
-import { Observable } from 'rxjs';
-import { MULTER_MODULE_ADAPTER, MULTER_MODULE_OPTIONS } from '../constants';
-import { MulterModuleOptions, MulterOptions } from '../interfaces';
-import { transformException } from '../utils';
-import {
-  CallHandler,
-  ExecutionContext,
-  Inject,
-  mixin,
-  NestInterceptor,
-  Optional,
-  Type,
-} from '@nestjs/common';
-
-type MulterInstance = any;
-
-export function FilesInterceptor(
-  fieldName: string,
-  maxCount?: number,
-  localOptions?: MulterOptions,
-): Type<NestInterceptor> {
-  class MixinInterceptor implements NestInterceptor {
-    protected multer: MulterInstance;
-
-    constructor(
-      @Optional()
-      @Inject(MULTER_MODULE_OPTIONS)
-      options: MulterModuleOptions = {},
-      @Inject(MULTER_MODULE_ADAPTER)
-      adapter: MulterInstance,
-    ) {
-      this.multer = (adapter as any)({
-        ...options,
-        ...localOptions,
-      });
-    }
-
-    async intercept(
-      context: ExecutionContext,
-      next: CallHandler,
-    ): Promise<Observable<any>> {
-      const ctx = context.switchToHttp();
-
-      await new Promise((resolve, reject) =>
-        this.multer.array(fieldName, maxCount)(
-          ctx.getRequest(),
-          ctx.getResponse(),
-          (err: any) => {
-            if (err) {
-              const error = transformException(err);
-              return reject(error);
-            }
-            resolve();
-          },
-        ),
-      );
-      return next.handle();
-    }
-  }
-  const Interceptor = mixin(MixinInterceptor);
-  return Interceptor as Type<NestInterceptor>;
-}
