--- packages/multer/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './files-upload-module.interface';
-export * from './multer-options.interface';
