--- packages/multer/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@nestjs/multer",
-  "version": "6.1.1",
-  "description": "...",
-  "author": "Kamil Mysliwiec",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/nestjs/nest"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {},
-  "peerDependencies": {
-    "@nestjs/common": "^6.0.0",
-    "@nestjs/core": "^6.0.0"
-  }
-}
