--- packages/multer/test/multer/fake-multer.ts
@@ -1,7 +1,0 @@
-export const fakeMulter = () => {
-  return {
-    array: () => {},
-    fields: () => {},
-    single: () => {},
-  };
-};
