--- packages/multer/test/multer/interceptors/file.interceptor.spec.ts
@@ -1,51 +1,0 @@
-import { CallHandler } from '@nestjs/common';
-import { ExecutionContextHost } from '@nestjs/core/helpers/execution-context-host';
-import { expect } from 'chai';
-import { of } from 'rxjs';
-import * as sinon from 'sinon';
-import { FileInterceptor } from '../../../interceptors/file.interceptor';
-import { fakeMulter } from '../fake-multer';
-
-describe('FileInterceptor', () => {
-  it('should return metatype with expected structure', async () => {
-    const targetClass = FileInterceptor('file');
-    expect(targetClass.prototype.intercept).to.not.be.undefined;
-  });
-  describe('intercept', () => {
-    let handler: CallHandler;
-    beforeEach(() => {
-      handler = {
-        handle: () => of('test'),
-      };
-    });
-
-    it('should call single() with expected params', async () => {
-      const fieldName = 'file';
-      const target = new (FileInterceptor(fieldName))(null, fakeMulter);
-
-      const callback = (req, res, next) => next();
-      const singleSpy = sinon
-        .stub((target as any).multer, 'single')
-        .returns(callback);
-
-      await target.intercept(new ExecutionContextHost([]), handler);
-
-      expect(singleSpy.called).to.be.true;
-      expect(singleSpy.calledWith(fieldName)).to.be.true;
-    });
-
-    it('should transform exception', async () => {
-      const fieldName = 'file';
-      const target = new (FileInterceptor(fieldName))(null, fakeMulter);
-      const err = {};
-      const callback = (req, res, next) => next(err);
-
-      (target as any).multer = {
-        single: () => callback,
-      };
-      (target.intercept(new ExecutionContextHost([]), handler) as any).catch(
-        error => expect(error).to.not.be.undefined,
-      );
-    });
-  });
-});
