--- packages/multer/test/multer/interceptors/files.interceptor.spec.ts
@@ -1,53 +1,0 @@
-import { CallHandler } from '@nestjs/common';
-import { ExecutionContextHost } from '@nestjs/core/helpers/execution-context-host';
-import { expect } from 'chai';
-import { of } from 'rxjs';
-import * as sinon from 'sinon';
-import { FilesInterceptor } from '../../../interceptors/files.interceptor';
-import { fakeMulter } from '../fake-multer';
-
-describe('FilesInterceptor', () => {
-  it('should return metatype with expected structure', async () => {
-    const targetClass = FilesInterceptor('file');
-    expect(targetClass.prototype.intercept).to.not.be.undefined;
-  });
-  describe('intercept', () => {
-    let handler: CallHandler;
-    beforeEach(() => {
-      handler = {
-        handle: () => of('test'),
-      };
-    });
-    it('should call array() with expected params', async () => {
-      const fieldName = 'file';
-      const maxCount = 10;
-      const target = new (FilesInterceptor(fieldName, maxCount))(
-        null,
-        fakeMulter,
-      );
-
-      const callback = (req, res, next) => next();
-      const arraySpy = sinon
-        .stub((target as any).multer, 'array')
-        .returns(callback);
-
-      await target.intercept(new ExecutionContextHost([]), handler);
-
-      expect(arraySpy.called).to.be.true;
-      expect(arraySpy.calledWith(fieldName, maxCount)).to.be.true;
-    });
-    it('should transform exception', async () => {
-      const fieldName = 'file';
-      const target = new (FilesInterceptor(fieldName))(null, fakeMulter);
-      const err = {};
-      const callback = (req, res, next) => next(err);
-
-      (target as any).multer = {
-        array: () => callback,
-      };
-      (target.intercept(new ExecutionContextHost([]), handler) as any).catch(
-        error => expect(error).to.not.be.undefined,
-      );
-    });
-  });
-});
