--- packages/multer/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./../tsconfig.base.json",
-  "compilerOptions": {
-    "baseUrl": "./"
-  },
-  "include": ["*.ts", "**/*.ts"]
-}
