+++ packages/platform-express/index.ts
@@ -7,4 +7,4 @@
 
 export * from './adapters';
 export * from './interfaces';
-export { default as multerAdapter } from 'multer';
+export * from './multer';
