+++ packages/platform-express/multer/files.constants.ts
@@ -1,0 +1,1 @@
+export const MULTER_MODULE_OPTIONS = 'MULTER_MODULE_OPTIONS';
