+++ packages/platform-express/multer/index.ts
@@ -1,0 +1,3 @@
+export * from './interceptors';
+export * from './interfaces';
+export * from './multer.module';
