+++ packages/platform-express/multer/interfaces/files-upload-module.interface.ts
