+++ packages/platform-express/multer/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './files-upload-module.interface';
