+++ packages/platform-express/multer/multer/multer.utils.ts
@@ -1,0 +1,24 @@
+import {
+  BadRequestException,
+  HttpException,
+  PayloadTooLargeException,
+} from '@nestjs/common';
+import { multerExceptions } from './multer.constants';
+
+export function transformException(error: Error | undefined) {
+  if (!error || error instanceof HttpException) {
+    return error;
+  }
+  switch (error.message) {
+    case multerExceptions.LIMIT_FILE_SIZE:
+      return new PayloadTooLargeException(error.message);
+    case multerExceptions.LIMIT_FILE_COUNT:
+    case multerExceptions.LIMIT_FIELD_KEY:
+    case multerExceptions.LIMIT_FIELD_VALUE:
+    case multerExceptions.LIMIT_FIELD_COUNT:
+    case multerExceptions.LIMIT_UNEXPECTED_FILE:
+    case multerExceptions.LIMIT_PART_COUNT:
+      return new BadRequestException(error.message);
+  }
+  return error;
+}
