+++ packages/platform-fastify/multer/interceptors/index.ts
@@ -1,0 +1,3 @@
+export * from './file-fields.interceptor';
+export * from './file.interceptor';
+export * from './files.interceptor';
