+++ packages/platform-fastify/multer/interfaces/files-upload-module.interface.ts
@@ -1,0 +1,19 @@
+import { Type } from '@nestjs/common';
+import { ModuleMetadata } from '@nestjs/common/interfaces';
+import { MulterOptions } from '../interfaces/multer-options.interface';
+
+export interface MulterModuleOptions extends MulterOptions {}
+
+export interface MulterOptionsFactory {
+  createMulterOptions(): Promise<MulterModuleOptions> | MulterModuleOptions;
+}
+
+export interface MulterModuleAsyncOptions
+  extends Pick<ModuleMetadata, 'imports'> {
+  useExisting?: Type<MulterOptionsFactory>;
+  useClass?: Type<MulterOptionsFactory>;
+  useFactory?: (
+    ...args: any[]
+  ) => Promise<MulterModuleOptions> | MulterModuleOptions;
+  inject?: any[];
+}
